Added random feature

+++ tests/php_server_kill.sh
@@ -1,2 +1,7 @@
 #!/bin/sh
+if ! command -v pkill &> /dev/null
+then
+    echo "pkill not found, server kill skipped"
+    exit
+fi
 pkill -e php
